requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
requests==2.32.3
beautifulsoup4==4.12.3
pandas==2.2.3
openpyxl==3.1.5
urllib3==2.2.2
